Networking refactor, finally beginning to see some okay code around here

+++ WindowStreamer Project/Shared/Constants.cs
@@ -6,7 +6,7 @@ namespace Shared
         public const int VideoStreamPort = 10064;
         public const int MetaStreamPort = 10063;
         public const int MetaFrameLength = 30;
-        public const float FramerateCap = 10f;
+        public const int FramerateCap = 10;
         public const char ParameterSeparator = ',';
         public const char SingleSeparator = '.';
